Move kbd shortcut to custom hook
